Refactor analytics code into folder.
