+++ .eslintignore
@@ -1,4 +1,4 @@
 coverage/*
 dist/
 node_modules/
-src/segment.js
+src/analytics/segment.js
