+++ config/webpack.common.config.js
@@ -4,7 +4,7 @@ const path = require('path');
 
 module.exports = {
   entry: {
-    segment: path.resolve(__dirname, '../src/segment.js'),
+    segment: path.resolve(__dirname, '../src/analytics/segment.js'),
     app: path.resolve(__dirname, '../src/index.jsx'),
   },
   output: {
