+++ src/components/App.jsx
@@ -8,7 +8,7 @@ import SiteFooter from '@edx/frontend-component-footer';
 import { fetchUserAccount, UserAccountApiService } from '@edx/frontend-auth';
 
 import apiClient from '../config/apiClient';
-import { handleTrackEvents } from '../analytics';
+import { handleTrackEvents } from '../analytics/analytics';
 import { getLocale, getMessages } from '../i18n/i18n-loader';
 import SiteHeader from './/SiteHeader';
 import ConnectedProfilePage from './ProfilePage';
